# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Milchvieh Datenanalyse-Assistent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /rules:
    get:
      tags: [Rules]
      operationId: listRules
      summary: List the current user's analysis rules
      responses:
        "200":
          description: Rules.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      tags: [Rules]
      operationId: createRule
      summary: Create an analysis rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleInput"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "400":
          $ref: "#/components/responses/BadRequest"
  /rules/{ruleId}:
    patch:
      tags: [Rules]
      operationId: updateRule
      summary: Update a rule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuleUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Rules]
      operationId: deleteRule
      summary: Delete a rule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
  /datasets/{datasetId}/warnings:
    get:
      tags: [Insights]
      operationId: listWarnings
      summary: List anomaly warnings for a dataset
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      responses:
        "200":
          description: Warnings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warning"
        "404":
          $ref: "#/components/responses/NotFound"
  /warnings/{warningId}:
    patch:
      tags: [Insights]
      operationId: updateWarning
      summary: Acknowledge or dismiss a warning
      parameters:
        - $ref: "#/components/parameters/WarningId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarningUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warning"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    DatasetId:
      name: datasetId
      in: path
      required: true
      schema:
        type: string
...
    RuleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
    WarningId:
      name: warningId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServerError:
      description: Server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...
    Rule:
      type: object
      required: [id, name, metric, enabled, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        metric:
          type: string
        comparator:
          type: ["string", "null"]
          enum: [gt, gte, lt, lte, eq, neq, null]
        threshold:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
          enum: [info, warning, critical, null]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    RuleInput:
      type: object
      required: [name, metric]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        metric:
          type: string
          minLength: 1
        comparator:
          type: string
          enum: [gt, gte, lt, lte, eq, neq]
        threshold:
          type: number
        unit:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        enabled:
          type: boolean
    RuleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        metric:
          type: string
          minLength: 1
        comparator:
          type: string
          enum: [gt, gte, lt, lte, eq, neq]
        threshold:
          type: number
        unit:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        enabled:
          type: boolean
    Warning:
      type: object
      required: [id, datasetId, title, severity, status, createdAt]
      properties:
        id:
          type: string
        datasetId:
          type: string
        title:
          type: string
        detail:
          type: ["string", "null"]
        metric:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        severity:
          type: string
          enum: [info, warning, critical]
        status:
          type: string
          enum: [open, acknowledged, dismissed]
        ruleId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    WarningUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, acknowledged, dismissed]
...